requirements.txt
absl-py==1.4.0
anyio==3.7.1
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
boto3==1.28.20
botocore==1.31.20
brevitas==0.8.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
coloredlogs==15.0.1
comm==0.1.4
concrete-ml==1.1.0
concrete-python==1.1.0rc1
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dependencies==2.0.1
dill==0.3.7
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.93.0
filelock==3.12.2
flatbuffers==2.0.7
fonttools==4.42.0
fsspec==2023.6.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
huggingface-hub==0.16.4
humanfriendly==10.0
hummingbird-ml==0.4.8
idna==3.4
ipykernel==6.25.0
ipython==8.14.0
jax==0.4.14
jedi==0.19.0
jmespath==1.0.1
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.13.1
onnxconverter-common==1.13.0
onnxmltools==1.11.0
onnxoptimizer==0.3.10
onnxruntime==1.13.1
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
scikit-learn==1.1.3
scipy==1.10.1
six==1.16.0
skl2onnx==1.12
skops==0.5.0
skorch==0.11.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.25.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
tf2onnx==1.14.0
threadpoolctl==3.2.0
torch==1.13.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.21.1
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.14.1
xgboost==1.6.2
